waiter/requirements.txt
Flask==1.1.2
requests==2.25.1
kafka-python==2.0.2
